Conflicts:
	languages/as3/src/com/github/tengi/client/TransportType.as
	languages/java/base/pom.xml
	languages/java/base/src/main/java/com/github/tengi/TransportType.java
	languages/java/base/src/main/java/com/github/tengi/buffer/ByteBufMemoryBuffer.java
	languages/java/base/src/main/java/com/github/tengi/buffer/MemoryBufferPool.java
	languages/java/base/src/main/java/com/github/tengi/transport/CompletionFutureAdapter.java
	languages/java/base/src/main/java/com/github/tengi/transport/WebsocketConnection.java
	languages/java/pom.xml
	languages/java/server/pom.xml
	languages/java/server/src/main/java/com/github/tengi/service/ServiceManager.java
	languages/java/server/src/main/java/com/github/tengi/transport/http/HttpConnection.java

+++ languages/java/base/src/main/java/com/github/tengi/buffer/MemoryBuffer.java
@@ -22,9 +22,9 @@ public interface MemoryBuffer
     extends ReadableMemoryBuffer, WritableMemoryBuffer
 {
 
-    long capacity();
+    int capacity();
 
-    long maxCapacity();
+    int maxCapacity();
 
     boolean growing();
