+++ languages/java/base/src/main/java/com/github/tengi/buffer/ReadableMemoryBuffer.java
@@ -1,5 +1,4 @@
 package com.github.tengi.buffer;
-
 /*
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
@@ -23,13 +22,12 @@ import com.github.tengi.UniqueId;
 
 import java.nio.ByteBuffer;
 
-
 public interface ReadableMemoryBuffer
 {
 
     boolean readable();
 
-    long readableBytes();
+    int readableBytes();
 
     int readBytes( byte[] bytes );
 
@@ -39,9 +37,9 @@ public interface ReadableMemoryBuffer
 
     int readBuffer( ByteBuffer byteBuffer, int offset, int length );
 
-    long readBuffer( WritableMemoryBuffer memoryBuffer );
+    int readBuffer( WritableMemoryBuffer memoryBuffer );
 
-    long readBuffer( WritableMemoryBuffer memoryBuffer, long offset, long length );
+    int readBuffer( WritableMemoryBuffer memoryBuffer, int offset, int length );
 
     boolean readBoolean();
 
@@ -69,8 +67,8 @@ public interface ReadableMemoryBuffer
 
     UniqueId readUniqueId();
 
-    long readerIndex();
+    int readerIndex();
 
-    void readerIndex( long readerIndex );
+    void readerIndex( int readerIndex );
 
 }
